chore: rever tsConfigLoader and configLoader back to dev, move compilerOptions check to tsconfig/index.ts
