+++ packages/remix-dev/compiler.ts
@@ -283,10 +283,14 @@ function isEntryPoint(config: RemixConfig, file: string) {
 
 ///////////////////////////////////////////////////////////////////////////////
 
+export interface FullBuildOptions extends Required<BuildOptions> {
+  incremental?: boolean;
+}
+
 async function buildEverything(
   config: RemixConfig,
   assetsManifestPromiseRef: AssetsManifestPromiseRef,
-  options: Required<BuildOptions> & { incremental?: boolean }
+  options: FullBuildOptions
 ): Promise<(esbuild.BuildResult | undefined)[]> {
   try {
     let browserBuildPromise = createBrowserBuild(config, options);
@@ -316,7 +320,7 @@ async function buildEverything(
 
 async function createBrowserBuild(
   config: RemixConfig,
-  options: BuildOptions & { incremental?: boolean }
+  options: FullBuildOptions
 ): Promise<esbuild.BuildResult> {
   // For the browser build, exclude node built-ins that don't have a
   // browser-safe alternative installed in node_modules. Nothing should
@@ -397,7 +401,7 @@ async function createBrowserBuild(
 
 function createServerBuild(
   config: RemixConfig,
-  options: Required<BuildOptions> & { incremental?: boolean },
+  options: FullBuildOptions,
   assetsManifestPromiseRef: AssetsManifestPromiseRef
 ): Promise<esbuild.BuildResult> {
   let dependencies = getAppDependencies(config);
